feat: allow passing config values as object move utils functions to its own file
