+++ package.json
@@ -68,6 +68,7 @@
     "go-ipfs-dep": "0.4.13",
     "guid": "0.0.12",
     "hapi": "^16.6.2",
+    "hat": "0.0.3",
     "http-browserify": "^1.7.0",
     "ipfs-api": "^17.2.0",
     "lodash.clonewith": "^4.5.0",
