+++ src/utils.js
@@ -1,0 +1,112 @@
+'use strict'
+
+const async = require('async')
+const fs = require('fs')
+const hat = require('hat')
+const os = require('os')
+const path = require('path')
+const exec = require('./exec')
+
+const join = path.join
+
+const isWindows = os.platform() === 'win32'
+
+// taken from https://github.com/hughsk/flat
+exports.flatten = (target) => {
+  let output = {}
+  const step = (object, prev) => {
+    object = object || {}
+    Object.keys(object).forEach(function (key) {
+      var value = object[key]
+      var isarray = Array.isArray(value)
+      var type = Object.prototype.toString.call(value)
+      var isbuffer = Buffer.isBuffer(value)
+      var isobject = (
+        type === '[object Object]' ||
+        type === '[object Array]'
+      )
+
+      var newKey = prev
+        ? prev + '.' + key
+        : key
+
+      if (!isarray && !isbuffer && isobject && Object.keys(value).length) {
+        return step(value, newKey)
+      }
+
+      output[newKey] = value
+    })
+  }
+
+  step(target)
+
+  return output
+}
+
+function tryJsonParse (input, callback) {
+  let res
+  try {
+    res = JSON.parse(input)
+  } catch (err) {
+    return callback(err)
+  }
+  callback(null, res)
+}
+
+exports.tryJsonParse = tryJsonParse
+
+// Consistent error handling
+exports.parseConfig = (path, callback) => {
+  async.waterfall([
+    (cb) => fs.readFile(join(path, 'config'), cb),
+    (file, cb) => tryJsonParse(file.toString(), cb)
+  ], callback)
+}
+
+exports.tempDir = (isJs) => {
+  return join(os.tmpdir(), `${isJs ? 'jsipfs' : 'ipfs'}_${hat()}`)
+}
+
+exports.findIpfsExecutable = (isJs, rootPath) => {
+  let appRoot = rootPath ? path.join(rootPath, '..') : process.cwd()
+  // If inside <appname>.asar try to load from .asar.unpacked
+  // this only works if asar was built with
+  // asar --unpack-dir=node_modules/go-ipfs-dep/* (not tested)
+  // or
+  // electron-packager ./ --asar.unpackDir=node_modules/go-ipfs-dep
+  if (appRoot.includes(`.asar${path.sep}`)) {
+    appRoot = appRoot.replace(`.asar${path.sep}`, `.asar.unpacked${path.sep}`)
+  }
+  const appName = isWindows ? 'ipfs.exe' : 'ipfs'
+  const depPath = isJs
+    ? path.join('ipfs', 'src', 'cli', 'bin.js')
+    : path.join('go-ipfs-dep', 'go-ipfs', appName)
+  const npm3Path = path.join(appRoot, '../', depPath)
+  const npm2Path = path.join(appRoot, 'node_modules', depPath)
+
+  if (fs.existsSync(npm3Path)) {
+    return npm3Path
+  }
+  if (fs.existsSync(npm2Path)) {
+    return npm2Path
+  }
+
+  throw new Error('Cannot find the IPFS executable')
+}
+
+function setConfigValue (node, key, value, callback) {
+  exec(
+    node.exec,
+    ['config', key, value, '--json'],
+    { env: node.env },
+    callback
+  )
+}
+
+exports.setConfigValue = setConfigValue
+
+exports.configureNode = (node, conf, callback) => {
+  async.eachOfSeries(conf, (value, key, cb) => {
+    setConfigValue(node, key, JSON.stringify(value), cb)
+  }, callback)
+}
